tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.fill.winding.evenodd.2.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.fill.winding.evenodd.2</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.fill.winding.evenodd.2</h1>
<p class="desc">evenodd winding number rule works in fill</p>


<script>
var t = async_test("evenodd winding number rule works in fill");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 100, 50);

  let path = new Path2D();
  path.rect(0, 0, 100, 50);
  path.rect(0, 0, 100, 50);
  path.closePath();

  ctx.fillStyle = "#f00";
  ctx.fill(path, "evenodd");

  _assertPixel(canvas, 50,25, 0,255,0,255);
  t.done();

});
</script>
